Consolidate other explainer tests into test_other.py
Following the project's test_<original_file>.py naming convention,
consolidated all 'other' explainer tests into a single file.

Contains 54 tests covering:
- TreeGain explainer (12 tests)
- Coefficient explainer (16 tests)
- Random explainer (13 tests)
- LimeTabular explainer (16 tests)

Test results: 28 passed, 26 skipped (optional dependencies)

Coverage improvements:
- TreeGain: 24% → 57%
- Coefficient: 62% → 100%
- Random: 29% → 79%
